+++ cw_ldk/lib/ffi.dart
@@ -31,20 +31,6 @@ typedef _error_message_utf8_Dart = int Function(
   int length,
 );
 
-/// <p class="para-brief"> remove later. another test for channels.</p>
-void ffi_channels(
-  Pointer<NativeFunction<Void Function(Pointer<ffi.Utf8>)>> func,
-) {
-  _ffi_channels(func);
-}
-final _ffi_channels_Dart _ffi_channels = _dl.lookupFunction<_ffi_channels_C, _ffi_channels_Dart>('ffi_channels');
-typedef _ffi_channels_C = Void Function(
-  Pointer<NativeFunction<Void Function(Pointer<ffi.Utf8>)>> func,
-);
-typedef _ffi_channels_Dart = void Function(
-  Pointer<NativeFunction<Void Function(Pointer<ffi.Utf8>)>> func,
-);
-
 /// C function `last_error_length`.
 int last_error_length() {
   return _last_error_length();
@@ -53,17 +39,26 @@ final _last_error_length_Dart _last_error_length = _dl.lookupFunction<_last_erro
 typedef _last_error_length_C = Int32 Function();
 typedef _last_error_length_Dart = int Function();
 
-/// <p class="para-brief"> remove later. test to see if channels work on phone.</p>
-void ldk_channels(
+/// C function `start_ldk`.
+Pointer<ffi.Utf8> start_ldk(
+  Pointer<ffi.Utf8> rpc_info,
+  Pointer<ffi.Utf8> ldk_storage_path,
+  Pointer<ffi.Utf8> mnemonic_key_phrase,
   Pointer<NativeFunction<Void Function(Pointer<ffi.Utf8>)>> func,
 ) {
-  _ldk_channels(func);
+  return _start_ldk(rpc_info, ldk_storage_path, mnemonic_key_phrase, func);
 }
-final _ldk_channels_Dart _ldk_channels = _dl.lookupFunction<_ldk_channels_C, _ldk_channels_Dart>('ldk_channels');
-typedef _ldk_channels_C = Void Function(
+final _start_ldk_Dart _start_ldk = _dl.lookupFunction<_start_ldk_C, _start_ldk_Dart>('start_ldk');
+typedef _start_ldk_C = Pointer<ffi.Utf8> Function(
+  Pointer<ffi.Utf8> rpc_info,
+  Pointer<ffi.Utf8> ldk_storage_path,
+  Pointer<ffi.Utf8> mnemonic_key_phrase,
   Pointer<NativeFunction<Void Function(Pointer<ffi.Utf8>)>> func,
 );
-typedef _ldk_channels_Dart = void Function(
+typedef _start_ldk_Dart = Pointer<ffi.Utf8> Function(
+  Pointer<ffi.Utf8> rpc_info,
+  Pointer<ffi.Utf8> ldk_storage_path,
+  Pointer<ffi.Utf8> mnemonic_key_phrase,
   Pointer<NativeFunction<Void Function(Pointer<ffi.Utf8>)>> func,
 );
 
@@ -79,41 +74,4 @@ typedef _store_dart_post_cobject_C = Void Function(
 );
 typedef _store_dart_post_cobject_Dart = void Function(
   Pointer<NativeFunction<Int8 Function(Int64, Pointer<Dart_CObject>)>> ptr,
-);
-
-/// <p class="para-brief"> Run LDK asynchronous</p>
-int test_ldk_async(
-  int isolate_port,
-  Pointer<ffi.Utf8> rpc_info,
-  Pointer<ffi.Utf8> ldk_storage_path,
-) {
-  return _test_ldk_async(isolate_port, rpc_info, ldk_storage_path);
-}
-final _test_ldk_async_Dart _test_ldk_async = _dl.lookupFunction<_test_ldk_async_C, _test_ldk_async_Dart>('test_ldk_async');
-typedef _test_ldk_async_C = Int32 Function(
-  Int64 isolate_port,
-  Pointer<ffi.Utf8> rpc_info,
-  Pointer<ffi.Utf8> ldk_storage_path,
-);
-typedef _test_ldk_async_Dart = int Function(
-  int isolate_port,
-  Pointer<ffi.Utf8> rpc_info,
-  Pointer<ffi.Utf8> ldk_storage_path,
-);
-
-/// <p class="para-brief"> remove later. for testing code that is blocking.</p>
-Pointer<ffi.Utf8> test_ldk_block(
-  Pointer<ffi.Utf8> path,
-  Pointer<NativeFunction<Void Function(Pointer<ffi.Utf8>)>> func,
-) {
-  return _test_ldk_block(path, func);
-}
-final _test_ldk_block_Dart _test_ldk_block = _dl.lookupFunction<_test_ldk_block_C, _test_ldk_block_Dart>('test_ldk_block');
-typedef _test_ldk_block_C = Pointer<ffi.Utf8> Function(
-  Pointer<ffi.Utf8> path,
-  Pointer<NativeFunction<Void Function(Pointer<ffi.Utf8>)>> func,
-);
-typedef _test_ldk_block_Dart = Pointer<ffi.Utf8> Function(
-  Pointer<ffi.Utf8> path,
-  Pointer<NativeFunction<Void Function(Pointer<ffi.Utf8>)>> func,
 );
