+++ cw_ldk/native/ldk-lib/Cargo.toml.old.toml
@@ -1,9 +1,31 @@
 [package]
 name = "ldk_lib"
 version = "0.1.0"
-edition = "2018"
+edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+# [dependencies]
+# lightning = { version = "0.0.106", features = ["max_level_trace"] }
+# lightning-block-sync = { version = "0.0.106", features = [ "rpc-client" ] }
+# lightning-invoice = { version = "0.13" }
+# lightning-net-tokio = { version = "0.0.106" }
+# lightning-persister = { version = "0.0.106" }
+# lightning-background-processor = { version = "0.0.106" }
+
+# base64 = "0.13.0"
+# bitcoin = "0.27"
+# bitcoin-bech32 = "0.12"
+# bech32 = "0.8"
+# hex = "0.3"
+
+# futures = "0.3"
+# chrono = "0.4"
+# rand = "0.4"
+# serde_json = { version = "1.0" }
+# tokio = { version = "1", features = [ "io-util", "macros", "rt", "rt-multi-thread", "sync", "net", "time" ] }
+
+# lazy_static = "1.4"
 
 [dependencies]
 lightning = { version = "0.0.106", features = ["max_level_trace"] }
@@ -25,8 +47,7 @@ rand = "0.4"
 serde_json = { version = "1.0" }
 tokio = { version = "1", features = [ "io-util", "macros", "rt", "rt-multi-thread", "sync", "net", "time" ] }
 
-[profile.release]
-panic = "abort"
+lazy_static = "1.4"
 
-[profile.dev]
-panic = "abort"
+[lib]
+doctest = false
