+++ cw_ldk/native/ldk-lib/src/bitcoind_client.rs
@@ -265,6 +265,7 @@ impl BroadcasterInterface for BitcoindClient {
 					if !err_str.contains("Transaction already in block chain")
 						&& !err_str.contains("Inputs missing or spent")
 						&& !err_str.contains("bad-txns-inputs-missingorspent")
+						&& !err_str.contains("txn-mempool-conflict")
 						&& !err_str.contains("non-BIP68-final")
 						&& !err_str.contains("insufficient fee, rejecting replacement ")
 					{
