+++ cw_ldk/native/ldk-lib/src/hex_utils.rs
@@ -1,6 +1,3 @@
-#![allow(dead_code)]
-#![allow(unused_variables)]
-
 use bitcoin::secp256k1::key::PublicKey;
 
 pub fn to_vec(hex: &str) -> Option<Vec<u8>> {
@@ -34,6 +31,9 @@ pub fn hex_str(value: &[u8]) -> String {
 }
 
 pub fn to_compressed_pubkey(hex: &str) -> Option<PublicKey> {
+	if hex.len() != 33 * 2 {
+		return None;
+	}
 	let data = match to_vec(&hex[0..33 * 2]) {
 		Some(bytes) => bytes,
 		None => return None,
